Rename recipe to recipe summary, swap URL for Spoonacular ID
The stored items only contain a title, image, and identifier —
no instructions — so "recipe" was a misnomer. Renamed across all
layers: API models, database table (migration), SQLDelight queries,
stores, routes, and client.

Replaced the url field on ApiRecipeSummary with a typed SpoonacularId
(Long) field so the record links back to its source unambiguously.
Updated ApiSaveRecipeSummary → ApiSaveSpoonacularRecipe accordingly.

All Spoonacular recipe-ID parameters in RecipeClient and its
implementations are now Long throughout, removing ad-hoc toInt()
conversions.

Added RecipeSummaryStoreTools so the find-recipe agent can save a
recipe summary for the user once they have chosen one.

+++ server/src/main/kotlin/se/gustavkarlsson/chefgpt/recipes/FakeRecipeClient.kt
@@ -33,7 +33,7 @@ class FakeRecipeClient : RecipeClient {
         """[{"id":641803,"title":"Easy & Delicious Berry Smoothie","image":"https://img.spoonacular.com/recipes/641803-312x231.jpg","imageType":"jpg","usedIngredientCount":2,"missedIngredientCount":1,"likes":12}]"""
 
     override suspend fun getRecipeInformation(
-        id: Int,
+        id: Long,
         includeNutrition: Boolean,
         addWinePairing: Boolean,
         addTasteData: Boolean,
@@ -41,13 +41,13 @@ class FakeRecipeClient : RecipeClient {
         """{"id":$id,"title":"Spaghetti Carbonara","readyInMinutes":30,"servings":4,"sourceUrl":"https://example.com/carbonara","image":"https://img.spoonacular.com/recipes/716429-556x370.jpg","summary":"A classic Italian pasta dish with eggs, cheese, pancetta, and pepper.","instructions":"<ol><li>Cook pasta.</li><li>Fry pancetta.</li><li>Mix eggs and cheese.</li><li>Combine and serve.</li></ol>","extendedIngredients":[{"id":1,"name":"spaghetti","amount":400.0,"unit":"g"},{"id":2,"name":"pancetta","amount":200.0,"unit":"g"},{"id":3,"name":"eggs","amount":4.0,"unit":""},{"id":4,"name":"parmesan cheese","amount":100.0,"unit":"g"},{"id":5,"name":"black pepper","amount":1.0,"unit":"tsp"}],"vegetarian":false,"vegan":false,"glutenFree":false,"dairyFree":false,"cuisines":["Italian"],"dishTypes":["lunch","main course","dinner"]}"""
 
     override suspend fun getRecipeInformationBulk(
-        ids: List<Int>,
+        ids: List<Long>,
         includeNutrition: Boolean,
     ): String =
         """[{"id":${ids.firstOrNull() ?: 1},"title":"Spaghetti Carbonara","readyInMinutes":30,"servings":4,"summary":"Classic Italian pasta.","vegetarian":false,"vegan":false}]"""
 
     override suspend fun getSimilarRecipes(
-        id: Int,
+        id: Long,
         number: Int,
     ): String =
         """[{"id":209128,"title":"Cacio e Pepe","readyInMinutes":25,"servings":2},{"id":227961,"title":"Pasta Aglio e Olio","readyInMinutes":20,"servings":4}]"""
@@ -64,28 +64,28 @@ class FakeRecipeClient : RecipeClient {
     ): String =
         """[{"id":716429,"title":"Chicken Parmesan"},{"id":715538,"title":"Chicken Tikka Masala"},{"id":716432,"title":"Chicken Alfredo"}]"""
 
-    override suspend fun summarizeRecipe(id: Int): String =
+    override suspend fun summarizeRecipe(id: Long): String =
         """{"id":$id,"title":"Spaghetti Carbonara","summary":"Spaghetti Carbonara is a classic Italian dish that takes about 30 minutes to prepare. It serves 4 and is perfect for a weeknight dinner. Each serving contains approximately 450 calories."}"""
 
     override suspend fun tasteById(
-        id: Int,
+        id: Long,
         normalize: Boolean,
     ): String =
         """{"sweetness":0.15,"saltiness":0.45,"sourness":0.05,"bitterness":0.02,"savoriness":0.85,"fattiness":0.65}"""
 
-    override suspend fun equipmentById(id: Int): String =
+    override suspend fun equipmentById(id: Long): String =
         """{"equipment":[{"name":"pot","image":"pot.png"},{"name":"frying pan","image":"pan.png"},{"name":"bowl","image":"bowl.jpg"}]}"""
 
-    override suspend fun priceBreakdownById(id: Int): String =
+    override suspend fun priceBreakdownById(id: Long): String =
         """{"ingredients":[{"name":"spaghetti","price":1.50},{"name":"pancetta","price":3.20},{"name":"eggs","price":0.80},{"name":"parmesan","price":2.50}],"totalCost":8.00,"totalCostPerServing":2.00}"""
 
     override suspend fun ingredientsById(
-        id: Int,
+        id: Long,
         measure: String,
     ): String =
         """{"ingredients":[{"name":"spaghetti","amount":{"metric":{"value":400.0,"unit":"g"}}},{"name":"pancetta","amount":{"metric":{"value":200.0,"unit":"g"}}},{"name":"eggs","amount":{"metric":{"value":4.0,"unit":""}}},{"name":"parmesan cheese","amount":{"metric":{"value":100.0,"unit":"g"}}}]}"""
 
-    override suspend fun nutritionById(id: Int): String =
+    override suspend fun nutritionById(id: Long): String =
         """{"calories":"450","carbs":"55g","fat":"18g","protein":"22g","nutrients":[{"name":"Calories","amount":450.0,"unit":"kcal","percentOfDailyNeeds":22.5},{"name":"Fat","amount":18.0,"unit":"g","percentOfDailyNeeds":27.7},{"name":"Protein","amount":22.0,"unit":"g","percentOfDailyNeeds":44.0},{"name":"Carbohydrates","amount":55.0,"unit":"g","percentOfDailyNeeds":18.3}]}"""
 
     override suspend fun getAnalyzedRecipeInstructions(instructions: String): String =
